Unique characteristics
Dynamic foldout was modified to accept additional filters for the class dropdown. The class FilteredDropdown, inherited from ClassDropDown, was created. In the bundle characteristics case, they are filtered by a property that defines them as unique or not. That way, it is not allowed to add more than 1 characteristic of the same type to a bundle if the characteristic is defined as unique.