on complete events are being stored correctly
Pending, try to break and find null references or broken cases. Also a pending bug when a node is added to the graph and the mouse is hovering a questaction, the GraphNodeChange action is duplicated, so all calls related to the selected quest node being changed are being called multiple times, not a priority but it will affect performance. Must fix eventually