Quest clean up and base method hook ready
Currently a scene game object can be selected multiple functions from the object's components can be called (could be scaled to have multiple objects) the issue is serialization and reference of that object, because the node data is not a monobehavior or a object, it can't store reference to scene gameobjects, as a solution the scene guid and the object name are stored and used to reload and find the object reference when the LBS tool loads its data. A better option would be to try and change the quest node data to be an object, but it could break multiple systems(mustcheck). A previous bug that caused the resuscription of an event that called the Quest Grammar Assistant task has been fixed